+++ CHANGELOG.md
@@ -23,7 +23,8 @@ CHANGELOG
  * Add config to set custom labels for GraphiQL JWT Authentication form fields
  * fix GraphiQL CORS error when use the explorer in a different domain or subdomain
  * [BC BREAK] Change definitions "extensions" to "plugins" to avoid confusion with CRUD extensions
-
+ * Add dataCollector to display helpful information in the web profiler
+ 
 1.0.6 - 18-03-23
 ----
  * Resolve array of IDs to real nodes